Creating the database
This time, we started creating the database that will be part of our social network. We created the main tables and fields that will be part of this ecosystem. Remember that we may make changes to the database, and with each new update to this project, it may be necessary to update the local database as well to avoid potential issues.