Creating more files
This time we went back to creating a few more files that will be part of our ecosystem. We created a 404.php page to handle navigation errors, and also a .htaccess file to manage page access on a developer level. Another thing we added was the header and footer structure of the site, where we’re going to create two separate files to be included and handled independently, leaving the page files responsible only for displaying the main content.