Creating function for validation
This time, we started creating the functions that will be responsible for performing the main user input validations. We need to ensure that both the email and the password provided by the user are actually valid. This validation must be done both on the client side and on the server side.