Changing structures of the main page
This time we started changing the structure of the main page to make it simpler and more presentable. We’re going to divide our page into three parts: on the left, there will be the menu and links to other pages; in the center, the main content with the posts; and on the right, other elements will be placed.